[FIXED] OpenCL target dim bug, can now draw gradient

+++ game/assets/cl/bake.cl
@@ -5,7 +5,7 @@ __kernel void render(write_only image2d_t out)
     const size_t width = get_global_size(0);
 	const size_t height = get_global_size(1);
     
-    write_imagef(out, (int2)(x, y), (float4)(x / width, y / height, 1.0, 1.0));
+    write_imagef(out, (int2)(x, y), (float4)(x / (float)(width), y / (float)(height), 1.0, 1.0));
 }
 
 // __kernel void render()
